logistics_admin/requirements.txt
Django
mysqlclient
django-cors-headers
djangorestframework-jwt
djangorestframework-simplejwt
djangorestframework
